+++ python/example/__main__.py
@@ -1,36 +1,15 @@
 """Entry point for package."""
 
-import argparse as ag
 import sys
 
-import example.sub_commands
+import example.cmdline
 
 
-def main(argv=None):
+def main():
     """
-    Start the Example tool.
-
-    :return: 0 on success.
+    Execute the CLI entry point.
     """
-    if not argv:
-        argv = sys.argv
-
-    # Parse arguments
-    parser = ag.ArgumentParser(prog=argv[0])
-    parser.add_argument('--version', action='version', version='%(prog)s ' + example.__version__)
-
-    subparser = parser.add_subparsers(title='Example Commands', description='Valid example commands.')
-    example.sub_commands.HelloCmd.init_subparser(subparser)
-
-    args = parser.parse_args(argv[1:])
-    try:
-        # Check if a sub-command is given, otherwise print help.
-        getattr(args, 'func')
-    except AttributeError:
-        parser.print_help()
-        return 2
-
-    return args.func(args)
+    return example.cmdline.main()
 
 
 if __name__ == '__main__':